requirements.txt
tensorflow
magenta
pyfluidsynth
streamlit
midi2audio
Pillow